fix a/+, a/# subscription
No need for the removed check, as it should be handled by mosquitto_sub_topic_check and mosquitto_pub_topic_check.
On the other hand it breaks a/b/+ against a/+/+ comparison, though perfectly acceptable.
At the same time a/b/+ is fine against a/# as comparison stops when sub reaches #.